# Optimize groupby.flatten