--- a/core/include/scipp/core/variable.h
+++ b/core/include/scipp/core/variable.h
@@ -964,7 +964,7 @@ namespace sparse {
 SCIPP_CORE_EXPORT Variable counts(const VariableConstProxy &var);
 SCIPP_CORE_EXPORT void reserve(const VariableProxy &sparse,
                                const VariableConstProxy &capacity);
-}
+} // namespace sparse
 
 } // namespace scipp::core