--- a/core/include/scipp/core/variable_view.h
+++ b/core/include/scipp/core/variable_view.h
@@ -51,8 +51,10 @@ public:
         m_targetDimensions(targetDimensions) {
     expectCanBroadcastFromTo(other.m_targetDimensions, m_targetDimensions);
     m_dimensions = other.m_dimensions;
+    // See implementation of ViewIndex regarding this relabeling.
     for (const auto label : m_dimensions.labels())
-      if (!other.m_targetDimensions.denseContains(label))
+      if (label != Dim::Invalid &&
+          !other.m_targetDimensions.denseContains(label))
         m_dimensions.relabel(m_dimensions.index(label), Dim::Invalid);
   }
 
@@ -71,8 +73,10 @@ public:
     m_dimensions = other.m_dimensions;
     if (begin != 0 || dim != Dim::Invalid)
       m_offset += begin * m_dimensions.offset(dim);
+    // See implementation of ViewIndex regarding this relabeling.
     for (const auto label : m_dimensions.labels())
-      if (!other.m_targetDimensions.denseContains(label))
+      if (label != Dim::Invalid &&
+          !other.m_targetDimensions.denseContains(label))
         m_dimensions.relabel(m_dimensions.index(label), Dim::Invalid);
   }