- Approx 3x performance improvement by calling `reserve` before concatenating vectors.
- Add `sparse::counts` to compute number of "counts". @reviewer: Is there a more generic name for this? Is it clear?

Benchmark results:

```
# params: n_spectrum/n_group
BM_groupby_flatten/64/1       521494478 ns    521477598 ns            1 bytes_per_second=2.85749G/s items_per_second=191.763M/s
BM_groupby_flatten/256/1      161263331 ns    161254483 ns            4 bytes_per_second=9.24077G/s items_per_second=620.138M/s
BM_groupby_flatten/1024/1     156758301 ns    156753283 ns            5 bytes_per_second=9.50612G/s items_per_second=637.945M/s
BM_groupby_flatten/4096/1     160004392 ns    159999957 ns            5 bytes_per_second=9.31323G/s items_per_second=625M/s
BM_groupby_flatten/16384/1    158559109 ns    158558848 ns            4 bytes_per_second=9.39787G/s items_per_second=630.681M/s
BM_groupby_flatten/65536/1    158034218 ns    158030470 ns            4 bytes_per_second=9.4293G/s items_per_second=632.789M/s
BM_groupby_flatten/262144/1   163133828 ns    163133136 ns            4 bytes_per_second=9.13436G/s items_per_second=612.996M/s
BM_groupby_flatten/64/4       205080564 ns    205069417 ns            9 bytes_per_second=7.2664G/s items_per_second=487.64M/s
BM_groupby_flatten/256/4      291085508 ns    291073534 ns            2 bytes_per_second=5.11938G/s items_per_second=343.556M/s
BM_groupby_flatten/1024/4     238472171 ns    238464389 ns            3 bytes_per_second=6.2488G/s items_per_second=419.35M/s
BM_groupby_flatten/4096/4     235461008 ns    235455625 ns            3 bytes_per_second=6.32865G/s items_per_second=424.708M/s
BM_groupby_flatten/16384/4    240695838 ns    240687515 ns            3 bytes_per_second=6.19108G/s items_per_second=415.476M/s
BM_groupby_flatten/65536/4    260194015 ns    260193008 ns            3 bytes_per_second=5.72696G/s items_per_second=384.33M/s
BM_groupby_flatten/262144/4   243282090 ns    243274173 ns            3 bytes_per_second=6.12525G/s items_per_second=411.059M/s
BM_groupby_flatten/64/16      138041546 ns    138036751 ns            5 bytes_per_second=10.7951G/s items_per_second=724.445M/s
BM_groupby_flatten/256/16     203240782 ns    203234408 ns            3 bytes_per_second=7.33201G/s items_per_second=492.043M/s
BM_groupby_flatten/1024/16    179154524 ns    179150284 ns            4 bytes_per_second=8.31769G/s items_per_second=558.191M/s
BM_groupby_flatten/4096/16    172914019 ns    172913319 ns            4 bytes_per_second=8.61771G/s items_per_second=578.324M/s
BM_groupby_flatten/16384/16   176010177 ns    176009250 ns            4 bytes_per_second=8.46612G/s items_per_second=568.152M/s
BM_groupby_flatten/65536/16   182268314 ns    182264936 ns            4 bytes_per_second=8.17555G/s items_per_second=548.652M/s
BM_groupby_flatten/262144/16  181999312 ns    181994955 ns            4 bytes_per_second=8.18768G/s items_per_second=549.466M/s
BM_groupby_flatten/64/64       82731193 ns     82731004 ns            8 bytes_per_second=18.0116G/s items_per_second=1.20874G/s
BM_groupby_flatten/256/64     168830971 ns    168827917 ns            4 bytes_per_second=8.82624G/s items_per_second=592.319M/s
BM_groupby_flatten/1024/64    165653145 ns    165652751 ns            4 bytes_per_second=8.99542G/s items_per_second=603.672M/s
BM_groupby_flatten/4096/64    157389086 ns    157386826 ns            4 bytes_per_second=9.46786G/s items_per_second=635.377M/s
BM_groupby_flatten/16384/64   161211320 ns    161211197 ns            4 bytes_per_second=9.24325G/s items_per_second=620.304M/s
BM_groupby_flatten/65536/64   174446597 ns    174443991 ns            4 bytes_per_second=8.54209G/s items_per_second=573.25M/s
BM_groupby_flatten/262144/64  163086419 ns    163083055 ns            4 bytes_per_second=9.13716G/s items_per_second=613.184M/s
```
This is essentialy equivalent to TOF-only event data.